+++ internal/core/router_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	"daily-help/internal/wecom"
+	"github.com/zcw199604/wecom-home-ops/internal/wecom"
 )
 
 type recordWeCom struct {