+++ internal/qinglong/provider.go
@@ -9,8 +9,8 @@ import (
 	"strings"
 	"time"
 
-	"daily-help/internal/core"
-	"daily-help/internal/wecom"
+	"github.com/zcw199604/wecom-home-ops/internal/core"
+	"github.com/zcw199604/wecom-home-ops/internal/wecom"
 )
 
 type Instance struct {